# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/department/{id}": {
      "get": {
        "operationId": "DepartmentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDepartmentDto"
                }
              }
            }
          }
        },
        "tags": [
          "department"
        ]
      },
      "patch": {
        "operationId": "DepartmentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDepartmentDto"
                }
              }
            }
          }
        },
        "tags": [
          "department"
        ]
      },
      "delete": {
        "operationId": "DepartmentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDepartmentDto"
                }
              }
            }
          }
        },
        "tags": [
          "department"
        ]
      }
    },
    "/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRespondeDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/student": {
      "post": {
        "operationId": "UserController_createStudent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/teacher": {
      "post": {
        "operationId": "TeacherController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTeacherDto"
                }
              }
            }
          }
        },
        "tags": [
          "teacher"
        ]
      },
      "get": {
        "operationId": "TeacherController_findAll",
        "parameters": [
          {
            "name": "schoolId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTeacherDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "teacher"
        ]
      }
    },
    "/teacher/department/{id}": {
      "get": {
        "operationId": "TeacherController_findAllDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTeacherDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "teacher"
        ]
      }
    },
    "/teacher/{id}": {
      "get": {
        "operationId": "TeacherController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTeacherDto"
                }
              }
            }
          }
        },
        "tags": [
          "teacher"
        ]
      },
      "patch": {
        "operationId": "TeacherController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTeacherDto"
                }
              }
            }
          }
        },
        "tags": [
          "teacher"
        ]
      },
      "delete": {
        "operationId": "TeacherController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTeacherDto"
                }
              }
            }
          }
        },
        "tags": [
          "teacher"
        ]
      }
    },
...
  },
  "info": {
    "title": "SIGEIT-API Documentation",
    "description": "API completa para el Sistema de Gestión de Información Educativa y Tecnológica",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "IdCreateEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "Department": {
        "type": "object",
        "properties": {}
      },
...
      "School": {
        "type": "object",
        "properties": {}
      },
      "ResponseDepartmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "schoolId": {
            "type": "number"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "abbreviation",
          "schoolId",
          "school",
          "status"
        ]
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "school": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "teacher": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          },
          "school": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "career": {
            "$ref": "#/components/schemas/IdCreateEntity"
          }
        },
        "required": [
          "email",
          "role",
          "idDocument",
          "name",
          "teacher",
          "status",
          "school",
          "department",
          "career"
        ]
      },
      "Teacher": {
        "type": "object",
        "properties": {}
      },
      "Career": {
        "type": "object",
        "properties": {}
      },
      "UserRespondeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "teacherId": {
            "type": "number"
          },
          "teacher": {
            "$ref": "#/components/schemas/Teacher"
          },
          "schoolId": {
            "type": "number"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "departmentId": {
            "type": "number"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "careerId": {
            "type": "number"
          },
          "career": {
            "$ref": "#/components/schemas/Career"
          }
        },
        "required": [
          "id",
          "name",
          "lastName",
          "firstName",
          "email",
          "role",
          "idDocument",
          "status",
          "teacherId",
          "teacher",
          "schoolId",
          "school",
          "departmentId",
          "department",
          "careerId",
          "career"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "teacher": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          },
          "school": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "career": {
            "$ref": "#/components/schemas/IdCreateEntity"
          }
        }
      },
      "CreateTeacherDto": {
        "type": "object",
        "properties": {
          "idDocument": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "idDocument",
          "firstName",
          "department",
          "status"
        ]
      },
      "ResponseTeacherDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "idDocument": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "departmentId": {
            "type": "number"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "idDocument",
          "firstName",
          "departmentId",
          "department",
          "status"
        ]
      },
      "UpdateTeacherDto": {
        "type": "object",
        "properties": {
          "idDocument": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}